updates: - added protected routes - added hover on shift table - some styling adjustment
