+++ reactabsensirsudsamratadmin/src/index.js
@@ -1,18 +1,19 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import "reactjs-popup/dist/index.css";
-import "react-calendar/dist/Calendar.css";
-import { Provider } from "react-redux";
-import store from "./config/authState/store";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import { router } from './App';
+import reportWebVitals from './reportWebVitals';
+import 'reactjs-popup/dist/index.css';
+import 'react-calendar/dist/Calendar.css';
+import { Provider } from 'react-redux';
+import store from './config/authState/store';
+import { RouterProvider } from 'react-router-dom';
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <RouterProvider router={router} />
     </Provider>
   </React.StrictMode>
 );
